Revert "haskell.packages: remove constraint on monoid-extras"
This reverts commit 52d69816b0f86c67844bfa34c22e86cdc187d074.

Unfortunately there is no way to update to monoid-extras 0.6 yet without
marking some packages as broken. The issue is that not all diagrams*
packages have had an update adding support for GHC 9.x yet (which would
also include monoid-extras 0.6 support). The only alternative to pinning
diagrams* and monoid-extras would be to have mismatched versions between
them which always causes issues with haskellPackages.diagrams.

Note that this commit re-introduces some build failures which are to be
fixed in a follow-up commit.