mercurialFull: add support for experimental in-tree git extension
The git extension is bundled in-tree, similarly to the hg gui, but also
requires pygit2. Enable it for mercurialFull, but don't bother with it
for the stock mercurial install.

Alas, it doesn't (yet?) work on nixpkgs because it doesn't support
octopus merges:

mercurial.error.ProgrammingError: git support can't handle octopus
merges, found a commit with 16 parents :(

I'm sure that commit is loved equally by all 16 parents.