gnomeExtensions.material-shell: 12 -> 40.a
Gnome has been updated to 40 and that required all extensions to be
changed to be made compatible. Version 12 of material-shell is not
compatible with gnome 40 so it does not work with current version of
gnome.
This updates material-shell to version 40.a which is compatible with
gnome 40.
Since version 12 material-shell has been written in typescript so it
needs to be compiled. Therefor I pull in all the development
dependencies.