python3: set an explicit high priority
Currently there are multiple derivations named python3, none of which have an
explicit meta.priority. This means that nix-env -i python3 will give a
different result than nix-env -iA nixpkgs.python3, and nix-env -u is not useful
when you have nixpkgs.python3 installed.

So we explicitly raise the priority of nixpkgs.python3, so that nix-env -i
python3 will no longer try to install an alpha version of python3 just because
it has a higher version number.