git: Drop POSIXPERM for the tests
Shared permissions are forbidden in the Nix build sandbox which is why
some of the tests fail. Dropping the POSIXPERM test prerequisite, that
is added unconditionally on Linux, should help us avoiding uninteresting
test failures after Git updates. The drawback is that this disables a
few additional tests (as of Git 2.32.0 approx. 71 out of more than 20k
subtests require POSIXPERM).