cvs-fast-export: clean up; fix cross
Style was unidiomatic for modern Nixpkgs, there was no
nativeBuildInputs/buildInputs separation, and some dependencies were
unused.

I've removed doCheck here, because the default Nixpkgs checkPhase
wasn't actually running the tests (need to cd into tests first), and
when I did write a checkPhase to run the tests, they didn't work and I
couldn't make them work.

A note on RCS:  There is code in cvsconvert that calls RCS, but I
analyzed it quite carefully, and came to the conclusion that all such
code is unreachable, so I removed the RCS dependency.