harfbuzz: Remove cairo dependency
Upstream recommends cairo, but it is only used for development
purposes and is not part of the library.
This removes transitive (build) dependencies on various X11 or
other GUI-related libraries, which is desirable for minimizing
rebuilds and maximizing parallellism.

This also adds fontconfig to buildInputs, because this direct lib
dependency was previously found via cairo's propagatedBuildInputs.