kodiPackages: make the package set extensible
With this change we gain the ability to create a kodiPackages set for a
given kodi instance that can be extended in user configurations.

Downstream consumers can now declare an overlay to add another package
(or change a dependency like so):

```nix
{
  nixpkgs.overlays = [
    (self: super: {
      kodiPackages = super.kodiPackages.extend (kself: ksuper: {
      	somePackage = ksuper.override {
          someFeatureFlag = true;
	};
      });
    })
  ];
}
```

The `kodePackages` attribute is equivalent to the previous package set
for the kodi package. The other package sets for kodi did previosly
exist but weren't accessible to users. They only ever surfaced through
the `withPackages` function.