imv: clarify license
licenses.gpl2 is deprecated, either gpl2Only or gpl2Plus should be
specified. However, imv is dual-licensed under the GPL and MIT licenses.

MIT is the one in the root LICENSE file, it's the one that Github shows
in  the main repo page and is also the more permissive one, therefore I
changed it to licenses.mit.