glib: add removeGIOModuleCache postIntall phase
glib-querymodules may leave behind a giomodule.cache.
However, this cache does not make sense on a per-package
basis, as it's meant to represent the present modules
in a given environment. So when many packages are introduced
these cache will no longer be valid. Not to mention they
cause collisions when present with other packages.