libvpx: fix cross-compilation
This is done by removing some warnings that are not supported by the
cross-compiler.

We also need to disable the --enable-external-build flag, which breaks
the build completely.
This PR: https://github.com/NixOS/nixpkgs/pull/100210 added the flag,
but it doesn't explain why it was added. It also doesn't show any
attempt at trying to cross-compile for either darwin or aarch64. So
let's remove it again, since it 'just' seems to break the build anyway.

- drop commented out configure flags
- be (more) explicit in patchShebangs
- libvpx_1_8: be explicit about patchShebangs, remove commented flags