pkgs/top-level/release.nix: include ocamlPackages in nixpkgs:trunk again
ocamlPackages has been removed in the past due to build capacity
shortage, this situation is not today's anymore, so we can remove this
“temporary” workaround.

The stress caused by this shouldn't be too bad — the whole ocamlPackages
set builds under one hour on normal hardware.

I'll leave it to the respective ecosystem's maintainers or other
concerned parties to re-add the remaining package sets like perlPackages
or rPackages.