mautrix-whatsapp: 0.1.6 -> unstable-2021-06-15
On `master` there are a few more bugfixes. Most notably, each message
sent by me was marked as "not sent" by WhatsApp for the last four days
until you hit a "Resend" in WhatsApp.

As the update to the latest `master` has solved the issue and being able
to correctly send messages is one of the core features of this package,
I decided to update the package in `nixpkgs` as well.