commanderx16: create x16-run script
In a typical FHS-compliant system, the rom and the emulator share the same
directory tree, and x16emu explores it. But Nix/Nixpkgs breaks this assumption.

Therefore, I made a useful script tailored to a regular user of Nixpkgs invoke
it in a more straightforward way. With this, the end user does not need to
explicitly set the default rom path.