lib/customisation: Use lists of attrsets for callPackageWith
This commit allows `callPackage[s]With` to take a list of attrsets as the
autoArgs instead. Doing this allows for a very easy optimization: When
we want to `callPackageWith` two (or more) attribute sets, instead of
having to // them together, which has to allocate a _new_ attribute set
that can store both sides, we can just pass them as two separate
attribute sets instead. The callPackageWith then first intersects these
two attribute sets with the function arguments before merging them.
Because the number of function arguments is usually rather small, we
never allocate any new big attribute sets.

We essentially just use the fact that

  intersectAttrs funArgs (a // b) == intersectAttrs funArgs a // intersectAttrs funArgs b

to our advantage to not have to allocate huge attribute sets when a or
b are already huge.

In my measurements, this patch saves about up to 10% of memory usage
when evaluating packages!