xlockmore: add preConfigure to compile with DPMS support (#127051)
The build of xlockmore produced only the minimum set of features. DPMS
support, Xinerama etc was disabled because the changes in the postPatch
step were never applied by re-running autoconf.

Furthermore, DPMS was not supported because it couldn't find the
`dpms.h` header file in order to compile the support for DPMS features.

This patch uses the `autoreconfHook` in which we run
`autoconf` to apply the changes of `configure.ac` to the `configure`
script. The patch also uses the `libXext.dev` `buildInput` in order to
allow access to the right header files.