wlroots: 0.13.0 -> 0.14.0
The new release comes with breaking changes so we temporarily introduce
wlroots_0_13 for packages that don't yet support wlroots 0.14.
For the rest of the packages the required upstream patches for this new
wlroots release are fetched (if feasible).

(cherry picked from commit 203c8edcdac9491912cb21fa9d84392cb6a69eef)