vscode: fix source name
This commit fixes various long standing issues with the handling of
executable names in the source archive. Depending on whether you install
code or codium the executables are named differently. This commit let's
us customize the source exe name in the generic VSCode derivation, but
keeps the generated executableName intact.