valgrind: 3.16.1 -> 3.17.0
Version 3.17.0 contains support for sched_getattr and sched_setattr.
This fixes an issue where the following messages can be logged:

    WARNING: unhandled amd64-linux syscall 315

See: https://bugs.kde.org/show_bug.cgi?id=369029

In addition to the version bump, this commit enables doCheck (as the
tests now pass) and provides a separate pname and version.

The upstream 3.17.0 also reorganized the output, placing the
default.supp file in libexec rather than lib. The postInstall hook was
adjusted accordingly.

A patch file which has not been used since upgrade to 3.15.0 in 2019
was also removed.