CONTRIBUTING.md: Move to repo root, where it is more visible.
We found that many users found it difficult to locate this document.

Github supports it in the root, see:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

(cherry picked from commit 3c29ced243cc032806f680ae67845b9b3a879183)

This fixes the links on the nixos.org website, the stable branch
docs point to master on github, so both master and the release branch
need to have the file in the same location.

fixes https://github.com/NixOS/nixos-homepage/issues/736