nixos/qemu-vm: load nix db if boot via bootloader
Previously, we only loaded the host's shared DB when booting the kernel
directly, since this information was transmitted via the kernel command
line given to QEMU. We now also load this when booting via a bootloader.

This is implemented via recording the path to load in the /boot
partition, and then loading it via a systemd unit. Thus it is
independent of whether we choose (e.g.) systemd-boot or grub, as opposed
to the strategy of manually mangling the kernel parameters in the
bootloader configuration.