nixos/tests/home-assistant: don't test package
The modules overrides extraComponents which leads to a costly rebuild of
the home-assistant package with all tests.  Make it less costly by not
running the tests, as does the default for the package option.  The
package's tests are already run by ofborg on every pull request as well
as by Hydra when building home-assistant.