nixos/nixos-install: build system using local nix store first
We build the flake twice to populate the local nix store with
all of the paths needed to do installation.
This is to avoid situations where `nix-build --store <path> --extra-substitutres auto?trusted=1`
errors on paths not being found in the local nix store.

See https://github.com/NixOS/nixpkgs/issues/126141 for more details.