r-modules: fix edge case for R package import
The R package "import" has a name that clashes
with the nix function "import". The package
name in nixpkgs is correctly modified to
avoid the clash, but the name of the
R package was also being changed in
cran-packages.nix, which broke downloading
of the package.

cran-packages.nix has been manually patched,
and genreate-r-packages.R has been fixed so
future automatic updates will succeed.