arcan: create a wrapper
This commit creates a wrapper with the purpose of gracefully execute Arcan
applications (appls, in Arcan jargon).

In a typical FHS system, Arcan can be invoked by a commandline like `arcan
<appl>`, or more generally `arcan <arcan options> <appl> <appl options>`. In
order to emulate this behaviour, this wrapper was crafted.

It `symlinkJoin`s Arcan and the appls, and sets the relevant environment
variables accordingly.