pkgs/treewide: Remove ignored defaults of package arguments
The pattern of using `attr ? null` for optional dependencies is pointless,
because these attributes always exist in nixpkgs, meaning they override
the default value.

And for platforms that don't support some attribute, that is indicated
with an evaluation error, not with the attribute not existing.

So this commit removes all these patterns. This means that now only
attributes that _don't_ exist in nixpkgs have a default value, like
`enableGUI` or similar configuration flags, which can notably be changed
with `.override`.