lib.callPackageWith: Enforce that default arguments aren't overridden
This enforces the changes made in the parent commit: Package expressions
that use an argument with a default like `attr ? null` will now throw
an error when the `attr` attribute exists in nixpkgs, since it would
override the default.

This ensures that all defaults in package arguments are actually applied
and not unknowingly overridden by an attribute.