bmake: enable ksh test on non musl platforms
bmake runs the ksh test if /bin/ksh exists. This is never a good
indication on platforms where we can sandbox and causes problems on
darwin where this path may exist, but never would be in PATH.

We solve this problem by always enabling the test and adding ksh to
checkInputs.

ksh doesn't seem to compile with musl, so we disable it on that
platform.