mariadb: replace galera_new_cluster in postPatch instead of postInstall
This prevents build errors when building mariadb with systemd removed
from buildInputs, in which case the galera_new_cluster file isn't
copied to $out/bin.

Also fixes a small bug where the URL in the script became invalid
after replacing - now it correctly points to an article on the MariaDB
website.