firefox: add PGO support
PGO is currently Linux 64-bit only.

Updating Firefox requires some manual assistance which is unwanted. At
the moment the PGO data is fetched from Mozilla's CI and needs to have
its hash updated manually.

The revision for the PGO data is automatically fetched and updated
which when doing a typical update of Firefox. However, it involves
downloading the Firefox source which is slow. It is also sort of a hack
in that, it passes through rev and update-source-version allows you to
specify a revision.