foomatic-db-ppds (cups ppd files): init
`foomatic-db-ppds` uses the Foomatic database from the packages
`foomatic-db` and -- optionally -- `foomatic-db-nonfree`
and the perl modules from `foomatic-db-engine`
to generate about 8,800 ppd files.

The general structure of the build recipe is as follows:

* Merge `foomatic-db` and `foomatic-db-nonfree` into
  one package that represents the Foomatic database.
  The package `foomatic-db-nonfree` is optional
  as it taints the result license-wise;
  it will only be used if `withNonfreeDb`
  is to to `true` in the `callPackage` call.
  We create a tiny setup hook script that provides the combined
  database and sets an environment variable pointing to the
  database direcotry, which is expected by the foomatic engine.

* The final package's license and version are computed
  from the licenses and versions of the database packages.
  The license is set to `free` if each database-providing
  package has a free license, and to `unfree` otherwise.
  The version is simply the highest version
  of the database-providing packages.

* The final package uses `foomatic-compiledb`
  from the `foomatic-db-engine` package to extract
  all ppd files from the database packages.
  `patchPpdFilesHook` is used to patch most
  executable invocations in the ppd files
  so that they point to nix store paths.
  Finally, ppd files are gzipped to reduce storage
  (from about 550 MiB to 90 MiB installed).

The "nonfree" version of the package, i.e. the version that is
based on `foomatic-db-nonfree` in addition to `foomatic-db`,
contains about 120 additional ppd files
compared to the "free" version.
Since the "free" version already produces about 8,700
ppd files and hydra won't build the "nonfree" version,
the commit adds two package variables to `all-packages.nix`:

* `foomatic-db-ppds` is based on `foomatic-db` only
* `foomatic-db-ppds-withNonfreeDb`
  is also based on `foomaitc-db-nonfree`

The package introduced by this commit
is the result of combining other packages;
it is not the build product of a simple source tarball.
While it would also be possible to perform the ppd file
generation directly in the build process of the database
packages, this would yield further complexity as the
`foomatic-db-nonfree` package needs to be combined with the
`foomatic-db` package before ppd file extraction is possible.

There is no upstream product with a name that
could/should be used for the `name` attribute,
the variable name, or for the filename in nixpkgs.
Similar packages have different names across distributions:

* https://repology.org/projects/?search=openprinting
* https://repology.org/projects/?search=foomatic

The name `foomatic-db-ppds` seems to be most common
(albeit not really *that* common):

* https://repology.org/project/foomatic-db-ppds/versions

At least openSUSE splits their corresponding
package into multiple "binary" packages
(similar to our multi-output packages):

* https://build.opensuse.org/package/binaries/Printing/OpenPrintingPPDs/openSUSE_Tumbleweed

I considered something similar.
However, after doing some statistics,
I concluded that it's not worth the effort:
The biggest dependencies (`perl` and `cups-filters`) are
already present on most NixOS systems, and they cannot
be "split away" easily since it cannot be
done along a canonical line (e.g. printer driver).
Splitting directly by dependency risks that ppd files
unexpectedly "move from output to output" on package updates;
disappearing ppd files can be quite annoying for package users.